minor cmake reorganization
- update submodules recursively
- also rename Lin to Posix
