+++ CMakeModules/Git.cmake
@@ -1,3 +1,5 @@
+set(PROJECT_VERSION "unknown")
+
 find_package(Git)
 
 if(GIT_FOUND)
@@ -10,7 +12,7 @@ if(GIT_FOUND)
     set(PROJECT_VERSION ${GIT_VERSION})
   endif()
 
-  execute_process(COMMAND ${GIT_EXECUTABLE} submodule update --init
+  execute_process(COMMAND ${GIT_EXECUTABLE} submodule update --init --recursive
     WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
     OUTPUT_VARIABLE WOMM_VERSION
     OUTPUT_STRIP_TRAILING_WHITESPACE)
