+++ Doxyfile
@@ -983,7 +983,7 @@ RECURSIVE              = YES
 # Note that relative paths are relative to the directory from which doxygen is
 # run.
 
-EXCLUDE                = build docs lib out thirdparty Launcher
+EXCLUDE                = build docs lib out thirdparty src/Launcher
 
 # The EXCLUDE_SYMLINKS tag can be used to select whether or not files or
 # directories that are symbolic links (a Unix file system feature) are excluded
