+++ src/CMakeLists.txt
@@ -1,0 +1,11 @@
+cmake_minimum_required(VERSION 3.20.4 FATAL_ERROR)
+
+add_subdirectory(Application)
+add_subdirectory(Arcomage)
+add_subdirectory(Engine)
+add_subdirectory(GUI)
+add_subdirectory(Io)
+add_subdirectory(Launcher)
+add_subdirectory(Media)
+add_subdirectory(Platform)
+add_subdirectory(Utility)
