+++ src/Engine/CMakeLists.txt
@@ -64,14 +64,14 @@ set(ENGINE_HEADERS AssetsManager.h
 add_library(engine STATIC ${ENGINE_SOURCES} ${ENGINE_HEADERS})
 target_check_style(engine)
 
-target_link_libraries(engine game)
+target_link_libraries(engine game utility)
 target_compile_definitions(engine PRIVATE PROJECT_VERSION="${PROJECT_VERSION}"
                                           BUILD_COMPILER="${BUILD_COMPILER}"
                                           BUILD_PLATFORM="${BUILD_PLATFORM}"
                                           BUILD_TYPE="${BUILD_TYPE}")
 
 include_directories(${INCLUDE_DIRECTORIES} ${ZLIB_INCLUDE_DIRS})
-target_link_libraries(engine utility ${ZLIB_LIBRARIES})
+target_link_libraries(engine ${ZLIB_LIBRARIES})
 
 add_subdirectory(Graphics)
 add_subdirectory(Objects)
