+++ src/Media/CMakeLists.txt
@@ -10,14 +10,15 @@ set(MEDIA_HEADERS Media.h
                   Audio/AudioPlayer.h
                   Audio/OpenALSoundProvider.h)
 
-add_library(media STATIC ${MEDIA_SOURCES} ${MEDIA_HEADERS})
-
 if(APPLE)
   add_definitions(-DFF_API_NEXT -Wno-deprecated-declarations)
 endif()
 
+add_library(media STATIC ${MEDIA_SOURCES} ${MEDIA_HEADERS})
+target_link_libraries(media utility)
+
 include_directories(${INCLUDE_DIRECTORIES} ${OPENAL_INCLUDE_DIRS} ${FFMPEG_INCLUDE_DIRS})
-target_link_libraries(media utility ${OPENAL_LIBRARY} ${AVCODEC_LIBRARIES} ${AVFORMAT_LIBRARIES} ${AVUTIL_LIBRARIES} ${SWSCALE_LIBRARIES} ${SWRESAMPLE_LIBRARIES})
+target_link_libraries(media ${OPENAL_LIBRARY} ${AVCODEC_LIBRARIES} ${AVFORMAT_LIBRARIES} ${AVUTIL_LIBRARIES} ${SWSCALE_LIBRARIES} ${SWRESAMPLE_LIBRARIES})
 
 target_check_style(media)
